Check for the first card of a set only.
When checking for a set (eg 5 and 6 or 4 and 7) there is no point
checking for both of the members of the set on the stacks, since the
match will be made when checking for either one.  So only check for
the first member of each set.

This also more closely simulates how the game is played (covering the
lower member of each set first at each step.)  It doesn't make any
noticeable difference to the run time.

Signed-off-by: Russell Steicke <russells@adelie.cx>